pyproject.toml
[tool.poetry]
name = "Revelio"
version = "0.1.0"
description = ""
authors = ["Jack Henry"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
GitPython = "3.1.44"
Markdown = "3.7"
MarkupSafe = "3.0.2"
PyYAML = "6.0.2"
babel = "2.16.0"
beautifulsoup4 = "4.12.3"
certifi = "2024.12.14"
charset_normalizer = "3.4.1"
click = "8.1.8"
colorama = "0.4.6"
defusedxml = "0.7.1"
ghp_import = "2.1.0"
idna = "3.10"
jinja2 = "3.1.6"
mergedeep = "1.3.4"
mkdocs-auto-refresh-build-pages = "1.0.2"
mkdocs-backlinks = "0.9.1"
mkdocs-drawio-exporter = "0.10.2"
mkdocs-exporter = "6.2.0"
mkdocs-git-authors-plugin = "0.9.2"
mkdocs-git-revision-date-localized-plugin = "1.3.0"
mkdocs-kroki-plugin = "0.9.0"
mkdocs-material = "9.5.50"
mkdocs-open-in-new-tab = "1.0.8"
mkdocs = "1.6.1"
mkdocs_get_deps = "0.2.0"
mkdocs_material_extensions = "1.3.1"
packaging = "24.2"
paginate = "0.5.7"
pathspec = "0.12.1"
platformdirs = "4.3.6"
pygments = "2.19.1"
pymdown-extensions = "10.14.1"
python_dateutil = "2.9.0.post0"
pytz = "2024.2"
pyyaml_env_tag = "0.1"
regex = "2024.11.6"
requests = "2.32.3"
result = "0.17.0"
six = "1.17"
smmap = "5.0.2"
soupsieve = "2.6"
urllib3 = "2.3.0"
watchdog = "6.0.0"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
